URL conf factoring; mode-neutral views move to integrations/views.py
Phase 1 of #361. Structural refactor only — URL paths and behavior
change but the underlying functionality is identical.

The top-level `integrations/urls.py` previously imported views from
both `connector.views` and `importer.views` (aliased as plain
`views`) and intermingled them, masking which routes are capability-
specific vs. mode-neutral. Several views lived in `connector/views.py`
purely by historical accident — they were actually mode-neutral
(consumed by both Connect-sync and Import-run flows, or by the
attribute-edit mixin shared between them).

URL conf:
- New `connector/urls.py` owns 10 Connect-mode routes.
- New `importer/urls.py` owns 5 Import-mode routes.
- Top-level `urls.py` stripped to: 4 mode-neutral routes
  (placement, refine, attribute history/restore), two include()s
  for the capability sub-packages, and the per-service URL
  discovery loop.

URL paths now consistent:
- `/integrations/connect/...` for Connect-mode (was bare under
  `/integrations/...`).
- `/integrations/import/...` for Import-mode (unchanged).
- `/integrations/placement|refine|attribute/...` for mode-neutral.

URL names normalized to a uniform `integrations_connect_*` prefix
for all Connect-mode routes. Seven names renamed:
- integrations_select          -> integrations_connect_select
- integrations_disable         -> integrations_connect_disable
- integrations_pause           -> integrations_connect_pause
- integrations_resume          -> integrations_connect_resume
- integrations_health_status   -> integrations_connect_health_status
- integrations_pre_sync        -> integrations_connect_pre_sync
- integrations_sync            -> integrations_connect_sync
(Three were already correct: integrations_connect_home,
integrations_connect_configure, integrations_connect_manage.)

Mode-neutral views moved from `connector/views.py` to the existing
top-level `integrations/views.py` (which already hosts
`CapabilityConfigureView`):
- IntegrationPlacementView (with DISMISS_ACTION_VALUE and helpers)
- IntegrationRefineView
- IntegrationAttributeHistoryInlineView
- IntegrationAttributeRestoreInlineView

Reference sweep: 14 reverse() calls, 8 {% url %} tags, 9 imports
pruned from connector/views.py. The `SYNCHRONIZATION_LOCK_NAME`
string literal `'integrations_sync'` (used by both Connect-sync
and Import-run to serialize) was intentionally left alone —
it's a lock name, not a URL name, and changing it would require
data migration.
