+++ src/hi/integrations/templates/integrations/connector/modals/integrations_select.html
@@ -35,7 +35,7 @@ All Integrations
 
   <div class="col-3">
     {% if integration_data.integration.is_enabled %}
-    <a href="{% url 'integrations_disable' integration_id=integration_data.integration.integration_id %}"
+    <a href="{% url 'integrations_connect_disable' integration_id=integration_data.integration.integration_id %}"
        role="button" class="btn btn-danger btn-control" data-async="modal">
       {% icon "delete" size="sm" css_class="hi-icon-left" %}DISABLE
     </a>
