+++ src/hi/integrations/templates/integrations/connector/pages/integration_manage.html
@@ -6,7 +6,7 @@
   <nav class="nav flex-column hi-vertical-nav">
     <div class="hi-vertical-nav-header">
       <a class="btn btn-outline-primary w-100" role="button"
-         href="{% url 'integrations_select' %}"
+         href="{% url 'integrations_connect_select' %}"
          data-async="modal">
         {% icon "settings" size="sm" css_class="hi-icon-left" %}INTEGRATIONS
       </a>
@@ -59,7 +59,7 @@
           <div class="alert alert-info py-2 mt-2 mb-0" role="status">
             {% icon "sync" size="sm" css_class="hi-icon-left" %}{{ core.sync_check_result.summary_message }}
             Click
-            <a href="{% url 'integrations_pre_sync' integration_id=core.integration_data.integration_id %}"
+            <a href="{% url 'integrations_connect_pre_sync' integration_id=core.integration_data.integration_id %}"
                class="alert-link"
                data-async="modal">Update</a>.
           </div>
