+++ src/hi/integrations/templates/integrations/connector/pages/no_integrations.html
@@ -5,7 +5,7 @@
 </div>
 <div class="text-center m-4 p-4">
   <a class="btn btn-primary btn-lg" role="button"
-     href="{% url 'integrations_select' %}"
+     href="{% url 'integrations_connect_select' %}"
      data-async="modal">
     CONFIGURE INTEGRATIONS
   </a>
